chore: move the logging module out of the internal package
