+++ gnet.go
@@ -30,7 +30,7 @@ import (
 
 	"github.com/panjf2000/gnet/errors"
 	"github.com/panjf2000/gnet/internal"
-	"github.com/panjf2000/gnet/internal/logging"
+	"github.com/panjf2000/gnet/logging"
 )
 
 // Action is an action that occurs after the completion of an event.
