+++ internal/netpoll/epoll.go
@@ -32,8 +32,8 @@ import (
 	"golang.org/x/sys/unix"
 
 	"github.com/panjf2000/gnet/errors"
-	"github.com/panjf2000/gnet/internal/logging"
 	"github.com/panjf2000/gnet/internal/netpoll/queue"
+	"github.com/panjf2000/gnet/logging"
 )
 
 // Poller represents a poller which is in charge of monitoring file-descriptors.
