+++ listener_unix.go
@@ -28,11 +28,12 @@ import (
 	"sync"
 	"time"
 
+	"golang.org/x/sys/unix"
+
 	"github.com/panjf2000/gnet/errors"
-	"github.com/panjf2000/gnet/internal/logging"
 	"github.com/panjf2000/gnet/internal/netpoll"
 	"github.com/panjf2000/gnet/internal/socket"
-	"golang.org/x/sys/unix"
+	"github.com/panjf2000/gnet/logging"
 )
 
 type listener struct {
