+++ listener_windows.go
@@ -26,8 +26,8 @@ import (
 	"sync"
 
 	"github.com/panjf2000/gnet/errors"
-	"github.com/panjf2000/gnet/internal/logging"
 	"github.com/panjf2000/gnet/internal/netpoll"
+	"github.com/panjf2000/gnet/logging"
 )
 
 type listener struct {
